re-org file structures
