+++ src/lib/Components/FilterModal.tsx
@@ -1,8 +1,20 @@
-import { ArrowLeftIcon, ArrowRightIcon, Box, Button, CloseIcon, color, Flex, Sans, Serif } from "@artsy/palette"
+import { ArrowRightIcon, Box, Button, CloseIcon, color, Flex, Sans, Serif } from "@artsy/palette"
+import { SortOptionsScreen as SortOptions } from "lib/Components/ArtworkFilterOptions/SortOptions"
 import React from "react"
 import { FlatList, LayoutAnimation, Modal as RNModal, TouchableWithoutFeedback, ViewProperties } from "react-native"
 import NavigatorIOS from "react-native-navigator-ios"
 import styled from "styled-components/native"
+
+interface FilterModalProps extends ViewProperties {
+  closeModal?: () => void
+  navigator?: NavigatorIOS
+  isFilterArtworksModalVisible: boolean
+}
+
+interface FilterModalState {
+  isComponentMounted: boolean
+  sortableItems: Array<{ type: string; data: any }>
+}
 
 export class FilterModalNavigator extends React.Component<FilterModalProps, FilterModalState> {
   state = {
@@ -36,7 +48,6 @@ export class FilterModalNavigator extends React.Component<FilterModalProps, Filt
                 <Flex onTouchStart={this.props.closeModal} style={{ flexGrow: 1 }} />
                 <ModalInnerView visible={isComponentMounted}>
                   <NavigatorIOS
-                    ref="filterNav"
                     navigationBarHidden={true}
                     initialRoute={{
                       component: FilterOptions,
@@ -60,48 +71,14 @@ export class FilterModalNavigator extends React.Component<FilterModalProps, Filt
   }
 }
 
-class SortOptionsScreen extends React.Component<SortOptionsScreenProps> {
-  handleBackNavigation() {
-    this.props.navigator.pop()
-  }
-
-  renderSortOption = ({ item }) => {
-    return (
-      <ListItem>
-        <Flex p={2} flexDirection="row" justifyContent="space-between" flexGrow={1}>
-          <Serif size="3">{item}</Serif>
-        </Flex>
-      </ListItem>
-    )
-  }
-
-  render() {
-    return (
-      <Flex flexGrow={1}>
-        <SortHeader>
-          <Flex alignItems="flex-end" mt={0.5} mb={2}>
-            <Box ml={2} mt={2} onTouchStart={() => this.handleBackNavigation()}>
-              <ArrowLeftIcon fill="black100" />
-            </Box>
-          </Flex>
-          <Sans mt={2} weight="medium" size="4">
-            Sort
-          </Sans>
-          <Box></Box>
-        </SortHeader>
-        <Flex>
-          <FlatList
-            keyExtractor={(_item, index) => String(index)}
-            data={SortOptions}
-            renderItem={item => <Box>{this.renderSortOption(item)}</Box>}
-          />
-        </Flex>
-        <BackgroundFill />
-      </Flex>
-    )
-  }
+interface FilterOptionsState {
+  filterOptions: Array<{ type: string; data: any }>
 }
 
+interface FilterOptionsProps {
+  closeModal: () => void
+  navigator: NavigatorIOS
+}
 class FilterOptions extends React.Component<FilterOptionsProps, FilterOptionsState> {
   state = {
     filterOptions: [],
@@ -122,13 +99,13 @@ class FilterOptions extends React.Component<FilterOptionsProps, FilterOptionsSta
 
   handleNavigationToSortScreen = () => {
     this.props.navigator.push({
-      component: SortOptionsScreen,
+      component: SortOptions,
     })
   }
 
   renderFilterOption = ({ item: { type, onTap } }) => {
     return (
-      <ListItem>
+      <OptionListItem>
         <Flex p={2} flexDirection="row" justifyContent="space-between" flexGrow={1}>
           <Serif size="3">{type}</Serif>
           <Flex flexDirection="row" onTouchEnd={() => onTap()}>
@@ -138,7 +115,7 @@ class FilterOptions extends React.Component<FilterOptionsProps, FilterOptionsSta
             <ArrowRightIcon fill="black30" ml={0.3} mt={0.3} />
           </Flex>
         </Flex>
-      </ListItem>
+      </OptionListItem>
     )
   }
 
@@ -173,13 +150,7 @@ class FilterOptions extends React.Component<FilterOptionsProps, FilterOptionsSta
   }
 }
 
-const SortHeader = styled(Flex)`
-  flex-direction: row;
-  justify-content: space-between;
-  padding-right: 20px;
-`
-
-const BackgroundFill = styled(Flex)`
+export const BackgroundFill = styled(Flex)`
   background-color: ${color("black10")};
   flex-grow: 1;
 `
@@ -198,6 +169,16 @@ export const FilterArtworkButton = styled(Button)`
   width: 110px;
 `
 
+export const OptionListItem = styled(Flex)`
+  flex-direction: row;
+  justify-content: space-between;
+  border: solid 0.5px ${color("black10")};
+  border-right-width: 0;
+  border-left-width: 0;
+  flex: 1;
+  width: 100%;
+`
+
 const ModalBackgroundView = styled.View`
   background-color: #00000099;
   flex: 1;
@@ -214,56 +195,9 @@ const ModalInnerView = styled.View<{ visible: boolean }>`
   border-top-right-radius: 10px;
 `
 
-const ListItem = styled(Flex)`
-  flex-direction: row;
-  justify-content: space-between;
-  border: solid 0.5px ${color("black10")};
-  border-right-width: 0;
-  border-left-width: 0;
-  flex: 1;
-  width: 100%;
-`
-
-const SortOptions = [
-  "Default",
-  "Price (low to high)",
-  "Price (high to low)",
-  "Recently Updated",
-  "Recently Added",
-  "Artwork year (descending)",
-  "Artwork year (ascending)",
-]
-
-interface FilterModalProps extends ViewProperties {
-  closeModal?: () => void
-  navigator?: NavigatorIOS
-  isFilterArtworksModalVisible: boolean
-}
-
-interface FilterModalState {
-  isComponentMounted: boolean
-  sortableItems: Array<{ type: string; data: any }>
-}
-
-interface SortOptionsScreenProps {
-  navigator: NavigatorIOS
-}
-
-interface FilterOptionsState {
-  filterOptions: Array<{ type: string; data: any }>
-}
-
-interface FilterOptionsProps {
-  closeModal: () => void
-  navigator: NavigatorIOS
-}
-
 /**
  * TODOS:
- *
- * 3) Close modal on background click
  * 4) Add tests  (check out Bid Flow + Consignments flow)
  *  e.g. when you click sort the correct component renders in the nav stack
- * 5) Clean up all the old/unused code
  *
  */
